server/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
gunicorn==22.0.0
psycopg[binary]==3.2.3
